requirements.txt
streamlit
streamlit_cookies_manager
grpcio
pymongo
pandas
protobuf
pillow